requirements.txt
aiogram==3.13.1  # Версия может варьироваться, но это актуальная на момент написания
python-dotenv==1.0.1
mistralai==1.1.0